Add files via upload
Attrition is a problem that impacts all businesses, irrespective of geography, industry and size of the company. It is a major problem to an organization, and predicting turnover is at the forefront of the needs of Human Resources (HR) in many organizations.
Companies all over the world have been grappling with the challenges of retaining talent andmanaging loss through attrition, such as retirement or existing employees resigning fromtheir current positions.
Employee attrition can result in a potential loss of projects or workflow of a specific project at a company, causing an imbalance in human resource planning and a loss of team harmony and social purposes.